requirements.txt
maafw>=4.0.0
loguru
pillow
pytz
pandas